feat(hashtower): rename the copied files from Incremental Merkle Tree to HashTower
