+++ packages/hashtower/package.json
@@ -1,7 +1,7 @@
 {
-    "name": "@zk-kit/incremental-merkle-tree",
-    "version": "1.0.0",
-    "description": "Incremental Merkle tree implementation in TypeScript.",
+    "name": "@zk-kit/hashtower",
+    "version": "0.0.1",
+    "description": "HashTower implementation in TypeScript.",
     "license": "MIT",
     "iife": "dist/index.js",
     "unpkg": "dist/index.min.js",
@@ -18,17 +18,17 @@
         "README.md"
     ],
     "repository": "git@github.com:privacy-scaling-explorations/zk-kit.git",
-    "homepage": "https://github.com/privacy-scaling-explorations/zk-kit/tree/main/packages/incremental-merkle-tree",
+    "homepage": "https://github.com/privacy-scaling-explorations/zk-kit/tree/main/packages/hashtower",
     "author": {
-        "name": "Cedoor",
-        "email": "me@cedoor.dev",
-        "url": "https://cedoor.dev"
+        "name": "LCamel",
+        "email": "lcamel@gmail.com",
+        "url": "https://twitter.com/LCamel"
     },
     "scripts": {
         "build:watch": "rollup -c rollup.config.ts -w --configPlugin typescript",
         "build": "rimraf dist && rollup -c rollup.config.ts --configPlugin typescript",
         "prepublishOnly": "yarn build",
-        "docs": "typedoc src/index.ts --out ../../docs/incremental-merkle-tree"
+        "docs": "typedoc src/index.ts --out ../../docs/hashtower"
     },
     "publishConfig": {
         "access": "public"
